common/interval_set: return int64_t for size()

+++ src/include/btree_interval_set.h
@@ -239,7 +239,7 @@ class btree_interval_set {
     return _size == other._size && m == other.m;
   }
 
-  int size() const {
+  int64_t size() const {
     return _size;
   }
